+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
 
+## [0.0.30] - 2022-02-27
+
+### Added
+- Support for extra values
+- `name` attribute on `inform-field` for extra values.
+
+### Fixed
+- Some reset issues: passed values were not always remembered.
+
 ## [0.0.29] - 2022-02-27
 
 ### Fixed
-- method defaults to 'GET' when not specified
+- method defaults to 'GET' when not specified.
 
 ## [0.0.28] - 2022-02-27